split SymbolicDimMgr
